Use float for ncount to allow scientific notation (e.g. 1e8)

+++ trunk/tools/Python/mcrun/main.py
@@ -121,7 +121,7 @@ def add_mcstas_options(parser):
         help='set random seed (must be: SEED != 0)')
 
     add('-n', '--ncount',
-        metavar='COUNT', type=int, default=1000000,
+        metavar='COUNT', type=float, default=1000000,
         help='set number of neutrons to simulate')
 
     add('-t', '--trace',
